refactor(sink-mixedbread): extract the Mixedbread record-reconcile from search-core
New `sink-mixedbread` crate (packages/sink/mixedbread) owning the record-source
reconcile: sync_documents (upload new/changed, skip unchanged) + gc_documents
(delete stale), with their SyncReport/GcReport and the source_filter helper.
Built on search-core's Store abstraction and search_core::wait_until_indexed, so
it works against the production MixedbreadStore or the in-memory test store.

Completes the sink-* component scheme: the indexer now fans every source out to
`sink-mixedbread` (semantic) + `sink-parquet` (S3 parquet) symmetrically.
search-core keeps the code-sync (manifest/content-addressed) + query paths; it
no longer exports sync_documents/gc_documents/GcReport. Record-reconcile tests
moved with the code. Part of #503.
