+++ Cargo.toml
@@ -43,6 +43,7 @@ members = [
   "packages/source/meta",
   "packages/search-py",
   "packages/source/slack",
+  "packages/sink/mixedbread",
   "packages/sink/parquet",
   "packages/indexer",
   "packages/tap",
@@ -164,6 +165,7 @@ serde_json = "1"
 sha2 = "0.10"
 similar = "2"
 source-slack = { path = "packages/source/slack" }
+sink-mixedbread = { path = "packages/sink/mixedbread" }
 sink-parquet = { path = "packages/sink/parquet" }
 snafu = "0.9"
 strip-ansi-escapes = "0.2"
