+++ packages/sink/mixedbread/package.nix
@@ -1,0 +1,5 @@
+{
+  id = "sink-mixedbread";
+  inRustWorkspace = true;
+  passthruTests = true;
+}
